Moved things around a bit!
Actually, the only real change is moving /util to /server, opening /util up to things that make more sense as actual utility-type classes.
