+++ src/server/AdventureServerAPI.java
@@ -1,4 +1,4 @@
-package util;
+package server;
 
 import java.io.IOException;
 import java.net.*;
