+++ src/server/ConnectionListener.java
@@ -1,4 +1,4 @@
-package util;
+package server;
 
 public interface ConnectionListener {
 	public void handle ( ConnectionEvent e );
